requirements.txt
torch==1.9.0
networkx==2.5.1
rdkit-pypi==2021.03.05
torch_geometric==2.0.1
torch_scatter==2.0.8
torch_sparse==0.6.12
torch_cluster==1.5.9